requirements.txt
aiogram>=3.4.1
aiosqlite>=0.20.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
feedparser>=6.0.11
